test: Initial Cypress smoke test setup
