+++ package.json
@@ -15,10 +15,14 @@
     "e2e:ci": "ng e2e --configuration=ci --protractor-config=projects/storefrontapp-e2e/protractor.headless.conf.js",
     "e2e:cy:open": "yarn --cwd ./projects/storefrontapp-e2e-cypress run cy:open",
     "e2e:cy:run": "yarn --cwd ./projects/storefrontapp-e2e-cypress run cy:run",
+    "e2e:cy:run:smoke": "yarn --cwd ./projects/storefrontapp-e2e-cypress run cy:run:smoke",
     "e2e:cy:run:ci": "yarn --cwd ./projects/storefrontapp-e2e-cypress run cy:run --record --key $CYPRESS_KEY",
+    "e2e:cy:run:smoke:ci": "yarn --cwd ./projects/storefrontapp-e2e-cypress run cy:run:smoke --record --key $CYPRESS_KEY",
     "e2e:cy:start-open": "start-server-and-test start:ci http-get://localhost:4200 e2e:cy:open",
     "e2e:cy:start-run": "start-server-and-test start:ci http-get://localhost:4200 e2e:cy:run",
     "e2e:cy:start-run-ci": "start-server-and-test start:ci http-get://localhost:4200 e2e:cy:run:ci",
+    "e2e:cy:start-run-smoke": "start-server-and-test start:ci http-get://localhost:4200 e2e:cy:run:smoke",
+    "e2e:cy:start-run-smoke-ci": "start-server-and-test start:ci http-get://localhost:4200 e2e:cy:run:smoke:ci",
     "generate:changelog": "ts-node ./scripts/changelog.ts",
     "lint": "ng lint",
     "mockserver": "cd ./projects/backend/mockgenerator && npm run start",
