+++ projects/storefrontapp-e2e-cypress/cypress/integration/smoke/persistent-user.e2e-spec.ts
@@ -1,0 +1,11 @@
+context('Check login', () => {
+  before(() => {
+    cy.window().then(win => win.sessionStorage.clear());
+    cy.login('test-user-cypress@ydev.hybris.com', 'Password123.');
+    cy.visit('/');
+  });
+
+  it('should register successfully', () => {
+    cy.get('.cx-login-status__greet').should('contain', 'Test User');
+  });
+});
