+++ projects/storefrontapp-e2e-cypress/cypress/support/commands.ts
@@ -2,3 +2,4 @@ import 'cypress-testing-library/add-commands';
 import './ng-select.commands';
 import './select-user-menu-option.commands';
 import './require-logged-in.commands';
+import './login.commands';
