+++ projects/storefrontapp-e2e-cypress/cypress/support/require-logged-in.commands.ts
@@ -1,4 +1,6 @@
 import { generateMail, randomString } from '../helpers/user';
+import { config, login, setSessionData } from './utils/login';
+
 declare global {
   namespace Cypress {
     interface Chainable {
@@ -39,35 +41,6 @@ export interface RequireLoggedInDebugOptions {
 Cypress.Commands.add(
   'requireLoggedIn',
   (accountData?: AccountData, options: RequireLoggedInDebugOptions = {}) => {
-    const apiUrl = Cypress.env('API_URL');
-    const config = {
-      tokenUrl: `${apiUrl}/authorizationserver/oauth/token`,
-      newUserUrl: `${apiUrl}/rest/v2/electronics/users/?lang=en&curr=USD`,
-      client: {
-        client_id: Cypress.env('CLIENT_ID'),
-        client_secret: Cypress.env('CLIENT_SECRET')
-      }
-    };
-
-    function login(
-      uid: string,
-      password: string,
-      failOnStatusCode: boolean = true
-    ) {
-      return cy.request({
-        method: 'POST',
-        url: config.tokenUrl,
-        body: {
-          ...config.client,
-          grant_type: 'password',
-          username: uid,
-          password
-        },
-        form: true,
-        failOnStatusCode
-      });
-    }
-
     function loginAsGuest() {
       return cy.request({
         method: 'POST',
@@ -99,23 +72,6 @@ Cypress.Commands.add(
           Authorization: `bearer ${access_token}`
         }
       });
-    }
-
-    function setSessionData(data) {
-      const authData = {
-        userToken: {
-          token: data
-        },
-        clientToken: {
-          loading: false,
-          error: false,
-          success: false
-        }
-      };
-      cy.window().then(win => {
-        win.sessionStorage.setItem('auth', JSON.stringify(authData));
-      });
-      return data;
     }
 
     const defaultAccount = {
